requirements.txt
annotated-types==0.6.0
anyio==4.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
exceptiongroup==1.2.0
fastapi==0.109.2
h11==0.14.0
idna==3.6
iniconfig==2.0.0
packaging==23.2
pact-python==2.1.1
pluggy==1.4.0
psutil==5.9.8
pydantic==2.6.1
pydantic_core==2.16.2
pytest==8.0.0
requests==2.31.0
six==1.16.0
sniffio==1.3.0
starlette==0.36.3
tomli==2.0.1
typing_extensions==4.9.0
urllib3==2.2.0
uvicorn==0.27.0.post1